requirements.txt
Flask==2.1.2
kafka-python==2.0.2
elasticsearch==7.17.0
